big set of changes (read description)
updated support for more uploaded fonts: more color, reduced height of the general design/layout.

update the functionality of the % support tag. it now calculates support for the entire language (all required glyphs). Plus added a doc that details the implementation.

changed default color for the system fallback to light gray so it's easier to detect when the font stack is not supporting a language.